remove unnecessary firebase service locator
