+++ client_end_user/androidApp/src/androidMain/kotlin/com/beepbeep/di/NotificationModule.kt
@@ -1,8 +1,8 @@
 package com.beepbeep.di
 
-import com.beepbeep.notification.FCMNotificationImp
-import com.beepbeep.notification.FireBaseMsgServiceImpl
-import com.beepbeep.notification.IFCMNotification
+import com.beepbeep.notification.fcm.FCMNotificationImp
+import com.beepbeep.notification.firebaseMessaaging.FireBaseMsgServiceImpl
+import com.beepbeep.notification.fcm.IFCMNotification
 import data.gateway.service.IFireBaseMessageService
 import org.koin.android.ext.koin.androidContext
 import org.koin.core.module.dsl.bind
@@ -10,9 +10,6 @@ import org.koin.core.module.dsl.singleOf
 import org.koin.dsl.module
 
 val firebaseModule = module {
-//    single { FireBaseMsgServiceImpl() }
-//    single<IFireBaseMessageService> { FirebaseMessagingImpl() }
-
     singleOf(::FireBaseMsgServiceImpl) { bind<IFireBaseMessageService>() }
     single<IFCMNotification> { FCMNotificationImp(androidContext().applicationContext) }
 }
