--- client_end_user/androidApp/src/androidMain/kotlin/com/beepbeep/notification/ServiceLocator.kt
@@ -1,13 +1,0 @@
-package com.beepbeep.notification
-
-object ServiceLocator {
-    private var fcmNotification: IFCMNotification? = null
-
-    fun initialize(notificationListener: IFCMNotification) {
-        this.fcmNotification = notificationListener
-    }
-
-    fun getFCMNotification(): IFCMNotification? {
-        return fcmNotification
-    }
-}
