+++ client_end_user/androidApp/src/androidMain/kotlin/com/beepbeep/notification/fcm/IFCMNotification.kt
@@ -1,4 +1,4 @@
-package com.beepbeep.notification
+package com.beepbeep.notification.fcm
 
 import android.app.PendingIntent
